Let RAW api handle all types of error
